backend/requirements.txt
numpy~=2.0.0
questionary~=2.0.1
pillow~=10.4.0
